move more logic from PathManager to PathCollectionEntry
